feat: implement AOP-style rate limiting using custom annotations and HandlerInterceptor

+++ src/main/java/com/example/ratelimiter/entity/RateLimitBucket.java
@@ -19,7 +19,7 @@ public class RateLimitBucket {
     private String identifier;     //this will store the user's IP address
 
     private double tokens; // Current tokens in the bucket
-    private int maxCapacity; //Maximum Tokens (e.g. 10)
+    private double maxCapacity; //Maximum Tokens (e.g. 10)
     private double refillRate; //Tokens added per second(e.g., 10)
 
     private LocalDateTime lastRefillTime;   //The timestamp of the last request
